first successful algorithm for the calculator.
generating all possible variations of a state and adding them to a graph, then doing this for the new states.

we effectively treat this as a graph traversal problem where the shortest path represents the most simple balancer to turn one state into another.

relatively quick for easier inputs, but complexity scales fast so in the future we will want to add better pruning / path choice logic

also maybe clean this up so it doesn't look so terrible