Upgraded heuristic function to make better estimates.
Made path construction its own function to clean up the traversal loop

Reworked graph traversal to instead start from both the input and output modes and meet in the middle. this led to significant speed boosts

Reworked the splits and rev_splits functions to take in a state and produce the resulting states instead of just the values

Added function to validate if a state is worth exploring or complete garbage

Changed operation priorities for splitters and merges back to equal (one), and did the same for their reverse variants. this gave a slight speed boost.

Forgot the rest...